refactor(ui): promote dropdown scroll to base
The viewport-height cap and scrolling belong on the shared
DropdownMenuContent, not just ButtonDropdown: any long menu was
clipping on short screens. Bring the content in line with upstream
shadcn (max-h-(--radix-dropdown-menu-content-available-height) +
overflow-y-auto) so scrolling works for every dropdown.

Fold the chevron/fade affordance behind an opt-in scrollAffordance
prop, backed by a localized client sub-component
(dropdown-menu-scroll-area.tsx), so there is one component and one
mental model. Collapse the parallel ButtonDropdown/ folder back to a
single thin consumer that just sets scrollAffordance.
